Update Database to have a timestamp property
The timestamp property is then used to provide context for
when the request was initially made so that a Conflict
exception can be thrown if a Document was updated or
deleted after the initial request timestamp.

There were 3 options in how to add this timestamp:

1. Add the timestamp to the updateDocument() function
2. Add setTimestamp() and resetTimestamp()
3. Add a withRequestTimestamp() function

The problem with option 1 is that it would pollute the signature and
make it messy.

The problem with option 2 is it adds a global state that developers
would need to make sure to reset. Not doing so can lead to unexpected
behavior.

Option 3 will require modifying places were we want to use the request
timestamp, but this explicitness is best for maintainability.